Move tests into own directory
